requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiohttp-retry==2.9.1
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.34.0
comm==0.2.2
cryptography==43.0.3
decorator==5.1.1
Deprecated==1.2.14
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
drf-yasg==1.21.8
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
frozenlist==1.5.0
geojson==3.1.0
gmaps==0.9.0
googlemaps==4.10.0
idna==3.10
inflection==0.5.1
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.2
jupyterlab_widgets==3.0.13
matplotlib-inline==0.1.7
multidict==6.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.5
numpy==2.1.3
packaging==24.1
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
twilio==9.3.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
vonage==3.17.4
vonage-jwt==1.1.4
vonage-utils==1.1.4
wcwidth==0.2.13
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.17.1
